+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.18.0.0
 
+* Add `mkInlineDatum`, `mkHashedDatum`
 * Moved `wireSizeTxF` out of `EraTx`
 * Move to `DecCBOR` instance for `Block` to `testlib`
 * Removed constraints from `EraPParams` for both `PParamsHKD Identity era` and `PParamsHKD StrictMaybe era`: `EncCBOR`, `DecCBOR`, `ToCBOR`, `FromCBOR`, `ToJSON`, `FromJSON`